changed project structure. Separated model from configuration, security and utils packages
