--- src/main/java/com/online_store/entity/Role.java
@@ -1,12 +1,0 @@
-package com.online_store.entity;
-
-import org.springframework.security.core.GrantedAuthority;
-
-public enum Role implements GrantedAuthority {
-    USER, ADMIN;
-
-    @Override
-    public String getAuthority() {
-        return name();
-    }
-}
