--- src/main/java/com/online_store/repository/ProductRepository.java
@@ -1,11 +1,0 @@
-package com.online_store.repository;
-
-import com.online_store.entity.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-
-import java.util.Optional;
-
-public interface ProductRepository extends JpaRepository<Product, Long>, JpaSpecificationExecutor<Product> {
-    Optional<Product> getProductById(Long id);
-}
