--- src/main/java/com/online_store/repository/UserRepository.java
@@ -1,15 +1,0 @@
-package com.online_store.repository;
-
-import com.online_store.entity.User;
-import org.springframework.data.jpa.repository.EntityGraph;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    @EntityGraph(attributePaths = "roles")
-    Optional<User> findByEmail(String email);
-
-    @EntityGraph(attributePaths = "roles")
-    Boolean existsByEmail(String email);
-}
