--- src/main/java/com/online_store/security/services/UserDetailsImpl.java
@@ -1,61 +1,0 @@
-package com.online_store.security.services;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.online_store.entity.User;
-import lombok.Data;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.io.Serial;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.List;
-
-@Data
-public class UserDetailsImpl implements UserDetails {
-    @Serial
-    private static final long serialVersionUID = 1L;
-    private final Long id;
-    private final String email;
-    @JsonIgnore
-    private final String password;
-    private final boolean active;
-    private final Collection<? extends GrantedAuthority> authorities;
-
-    public static UserDetailsImpl create(User user) {
-        String userRole = user.getRoles().iterator().next().toString();
-        List<GrantedAuthority> authorities = Collections.singletonList(new SimpleGrantedAuthority(userRole));
-        return new UserDetailsImpl(user.getId(), user.getEmail(), user.getPassword(), user.getActive(), authorities);
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return active;
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return authorities;
-    }
-
-    @Override
-    public String getUsername() {
-        return email;
-    }
-}
