--- src/main/java/com/online_store/security/services/UserDetailsServiceImpl.java
@@ -1,30 +1,0 @@
-package com.online_store.security.services;
-
-import com.online_store.constants.ErrorMessage;
-import com.online_store.entity.User;
-import com.online_store.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.Optional;
-
-@Service("userDetailsServiceImpl")
-@RequiredArgsConstructor
-public class UserDetailsServiceImpl implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    @Override
-    @Transactional
-    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        Optional<User> user = userRepository.findByEmail(email);
-        if (user.isEmpty()) {
-            throw new UsernameNotFoundException(ErrorMessage.USER_NOT_FOUND);
-        }
-        return UserDetailsImpl.create(user.orElse(null));
-    }
-}
