+++ src/main/java/com/online_store/web/order/model/Order.java
@@ -1,0 +1,44 @@
+package com.online_store.web.order.model;
+
+import com.online_store.web.product.model.Product;
+import com.online_store.web.user.model.User;
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+
+@Entity
+@Table(name = "orders")
+@Getter
+@Setter
+public class Order {
+    @Id
+    @Column(name = "id")
+    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "orders_seq")
+    @SequenceGenerator(name = "orders_seq", sequenceName = "orders_seq", allocationSize = 1)
+    private Long id;
+
+    @Column(name = "total_price", nullable = false)
+    private Double totalPrice;
+
+    @Column(name = "created_at", columnDefinition = "timestamp default current_timestamp")
+    private LocalDateTime createdAt = LocalDateTime.now();
+
+    @Column(name = "status", nullable = false)
+    private String status;
+
+    @ManyToMany
+    @JoinTable(
+            name = "order_product",
+            joinColumns = @JoinColumn(name = "order_id"),
+            inverseJoinColumns = @JoinColumn(name = "product_id")
+    )
+    private List<Product> products = new ArrayList<>();
+
+    @ManyToOne
+    @JoinColumn(name = "user_id")
+    private User user;
+}
