+++ src/main/java/com/online_store/web/product/dto/SearchRequest.java
@@ -1,0 +1,14 @@
+package com.online_store.web.product.dto;
+
+import lombok.Data;
+import org.springframework.format.annotation.DateTimeFormat;
+
+import java.time.LocalDateTime;
+
+@Data
+public class SearchRequest {
+    private String city;
+    @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
+    private LocalDateTime date;
+    private Integer participantCount;
+}
