+++ src/main/java/com/online_store/web/product/service/CategoryService.java
@@ -1,0 +1,14 @@
+package com.online_store.web.product.service;
+
+import com.online_store.web.product.dto.CategoryDTO;
+import com.online_store.web.product.model.Category;
+
+import java.util.List;
+
+public interface CategoryService {
+    List<Category> getAllCategories();
+
+    Category getCategoryById(Long id);
+
+    Category createOrUpdateCategory(CategoryDTO existingCategory);
+}
