+++ src/main/java/com/online_store/web/user/dao/UserRepository.java
@@ -1,0 +1,15 @@
+package com.online_store.web.user.dao;
+
+import com.online_store.web.user.model.User;
+import org.springframework.data.jpa.repository.EntityGraph;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    @EntityGraph(attributePaths = "roles")
+    Optional<User> findByEmail(String email);
+
+    @EntityGraph(attributePaths = "roles")
+    Boolean existsByEmail(String email);
+}
