+++ src/main/java/com/online_store/web/user/model/Role.java
@@ -1,0 +1,12 @@
+package com.online_store.web.user.model;
+
+import org.springframework.security.core.GrantedAuthority;
+
+public enum Role implements GrantedAuthority {
+    USER, ADMIN;
+
+    @Override
+    public String getAuthority() {
+        return name();
+    }
+}
